#2361 AppFwk : Make sure field of FilePath supports space, add more tests

+++ Fwk/AppFwk/cafProjectDataModel/cafPdmXml/cafInternalPdmFilePathStreamOperators.cpp
@@ -18,7 +18,7 @@ QTextStream& operator<<(QTextStream& str, const std::vector<caf::FilePath>& file
         trimmedEntries.push_back(text);
     }
 
-    for (size_t i = 0; i < trimmedEntries.size(); i++)
+    for (int i = 0; i < trimmedEntries.size(); i++)
     {
         str << trimmedEntries[i];
